Move statedb
